Created methods from task 5 -> calculatingTableWhereProductCorrespondsWight Changed double to int in method getWeight

+++ src/base/products/Products.java
@@ -21,7 +21,7 @@ public class Products {
         this.productGrade = productGrade;
     }
 
-    public double getWeight() {
+    public int getWeight() {
         return weight;
     }
